backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
httpx==0.27.0
google-generativeai==0.8.0
pypdf==4.3.1
python-dotenv==1.0.1
aiofiles==24.1.0
jinja2==3.1.4
firecrawl-py==1.5.0
whisperlivekit
